Do contrast

Do contrast

Do contrast

Public fields

models

Model

contrasts

character with contrasts

contrast

function

modelName

model name

function

to adjust p-values (default LFQService::adjust_p_values) create Contrast

Active bindings

contrast

function

function

to adjust p-values (default LFQService::adjust_p_values) create Contrast

Methods

Public methods

Contrasts$new()

Contrasts$get_contrasts_sides()

Contrasts$get_linfct()

Contrasts$get_contrasts()

Contrasts$get_Plotter()

Contrasts$clone()

Method new()

Usage

Contrasts$new(
  models,
  contrasts,
  contrast_fun,
  subject_Id,
  modelName = "Model",
  p.adjust = LFQService::adjust_p_values
)

Arguments

models

a dataframe with a structure similar to that generated by build_model

contrasts

a character vector with contrast specificiation

contrast_fun

see. eg. my_contrast_V2,my_contest

subject_Id

columns with subject_Id (e.g. protein_Id)

p.adjust

function to adjust the p-values

Modelname

default "Model"

Method get_contrasts_sides()

get both sides of contrasts

Usage

Contrasts$get_contrasts_sides()

Method get_linfct()

get linear functions from contrasts

Usage

Contrasts$get_linfct()

Method get_contrasts()

get table with contrast estimates

Usage

Contrasts$get_contrasts(all = FALSE)

Arguments

all

should all columns be returned (default FALSE)

Returns

data.frame with contrasts

Method get_Plotter()

return Contrast_Plotter

Usage

Contrasts$get_Plotter()

Method clone()

The objects of this class are cloneable with this method.

Usage

Contrasts$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

rm(list = ls())
library(LFQService)
library(tidyverse)

D <- LFQService::ionstar$normalized()
D$data <- dplyr::filter(D$data ,protein_Id %in% sample(protein_Id, 100))
modelFunction <-
  make_custom_model_lmer("transformedIntensity  ~ dilution. + (1 | peptide_Id) + (1 | sampleName)")
pepIntensity <- D$data
config <- D$config
config$table$hkeysDepth()

#> [1] "protein_Id"

mod <- build_model(
 pepIntensity,
 modelFunction,
 subject_Id = config$table$hkeysDepth())

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> boundary (singular) fit: see ?isSingular

#> Joining, by = "protein_Id"

 Contr <- c("dil.b_vs_a" = "dilution.a - dilution.b")
 #LFQService::Contrasts$debug("get_linfct")
contrastX <- LFQService::Contrasts$new(mod$modelDF,
 Contr,
 modelFunction$contrast_fun,
 subject_Id = config$table$hkeysDepth(),
 modelName = modelFunction$model_name)

contrastX$get_contrasts_sides()

#> # A tibble: 1 x 3
#>   contrast   c1         c2        
#>   <chr>      <chr>      <chr>     
#> 1 dil.b_vs_a dilution.a dilution.b

contrastX$get_linfct()

#> dil.b_vs_a=dilution.a - dilution.b

#> $linfct
#>            (Intercept) dilution.b dilution.c dilution.d dilution.e
#> dilution.a           1          0          0          0          0
#> dilution.b           1          1          0          0          0
#> dilution.c           1          0          1          0          0
#> dilution.d           1          0          0          1          0
#> dilution.e           1          0          0          0          1
#> 
#> $linfct_A
#>            (Intercept) dilution.b dilution.c dilution.d dilution.e
#> dil.b_vs_a           0         -1          0          0          0
#> 

xx <- contrastX$get_contrasts()

#> dil.b_vs_a=dilution.a - dilution.b

#> Warning: Problem with `mutate()` input `contrast`.
#> i Model is rank deficient!
#> i Input `contrast` is `map(linear_model, contrastfun, linfct = linfct)`.
#> i The error occurred in group 25: protein_Id = "sp|P29745|PEPT_ECOLI~1689~P29745".

#> Warning: Model is rank deficient!

#> Joining, by = c("contrast", "protein_Id")

imputed <- get_imputed_contrasts(D$data, D$config, Contr)

#> completing cases

#> dil.b_vs_a=dilution.a - dilution.b