Ropeca

Ropeca

Ropeca

See also

Super classes

LFQService::Contrast -> LFQService::ContrastsModerated -> ContrastsROPECA

Methods

Public methods

ContrastsROPECA$get_contrast()

ContrastsROPECA$get_Plotter()

ContrastsROPECA$clone()

Inherited methods

LFQService::Contrast$get_contrasts()

LFQService::Contrast$get_contrasts_sides()

LFQService::Contrast$get_linfct()

LFQService::Contrast$initialize()

Method get_contrast()

get contrasts

Usage

ContrastsROPECA$get_contrast(all = FALSE)

Method get_Plotter()

Usage

ContrastsROPECA$get_Plotter()

Method clone()

The objects of this class are cloneable with this method.

Usage

ContrastsROPECA$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

D <- LFQService::ionstar$normalized()
D$data <- dplyr::filter(D$data ,protein_Id %in% sample(protein_Id, 100))
modelFunction <-
  make_custom_model_lm("transformedIntensity  ~ dilution.")
pepIntensity <- D$data
config <- D$config$clone(deep = TRUE)
config$table$hierarchyDepth <- 2
config$table$hkeysDepth()

#> [1] "protein_Id" "peptide_Id"

mod <- build_model(
 pepIntensity,
 modelFunction,
 subject_Id = config$table$hkeysDepth())

#> Warning: Problem with `mutate()` input `linear_model`.
#> i WARN :Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels
#> 
#> i Input `linear_model` is `purrr::map(data, modelFunction$model_fun, pb = pb)`.
#> i The error occurred in group 29: protein_Id = "CON__P02769~18~NA", peptide_Id = "VHKECCHGDLLECADDR".

#> Warning: WARN :Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels

#> Warning: Problem with `mutate()` input `linear_model`.
#> i WARN :Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels
#> 
#> i Input `linear_model` is `purrr::map(data, modelFunction$model_fun, pb = pb)`.
#> i The error occurred in group 241: protein_Id = "sp|P19367|HXK1_HUMAN~1483~P19367", peptide_Id = "EGLLFEGR".

#> Warning: WARN :Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels

#> Warning: Problem with `mutate()` input `linear_model`.
#> i WARN :Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels
#> 
#> i Input `linear_model` is `purrr::map(data, modelFunction$model_fun, pb = pb)`.
#> i The error occurred in group 604: protein_Id = "sp|Q03001|DYST_HUMAN~2575~Q03001", peptide_Id = "WQQTPAPTDTEAVK".

#> Warning: WARN :Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels

#> Warning: Problem with `mutate()` input `linear_model`.
#> i WARN :Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels
#> 
#> i Input `linear_model` is `purrr::map(data, modelFunction$model_fun, pb = pb)`.
#> i The error occurred in group 605: protein_Id = "sp|Q03001|DYST_HUMAN~2575~Q03001", peptide_Id = "YLTMLQTQQK".

#> Warning: WARN :Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels

#> Warning: Problem with `mutate()` input `linear_model`.
#> i WARN :Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels
#> 
#> i Input `linear_model` is `purrr::map(data, modelFunction$model_fun, pb = pb)`.
#> i The error occurred in group 707: protein_Id = "sp|Q99873|ANM1_HUMAN~3969~Q99873", peptide_Id = "VEDLTFTSPFCLQVK".

#> Warning: WARN :Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels

#> Warning: Problem with `mutate()` input `linear_model`.
#> i WARN :Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels
#> 
#> i Input `linear_model` is `purrr::map(data, modelFunction$model_fun, pb = pb)`.
#> i The error occurred in group 711: protein_Id = "sp|Q9BY44|EIF2A_HUMAN~4119~Q9BY44", peptide_Id = "ATIFNLK".

#> Warning: WARN :Error in `contrasts<-`(`*tmp*`, value = contr.funs[1 + isOF[nn]]): contrasts can be applied only to factors with 2 or more levels

#> Joining, by = c("protein_Id", "peptide_Id")

 Contr <- c("dil.b_vs_a" = "dilution.a - dilution.b")
 #ContrastROPECA$debug("get_contrast")
 contrast <- LFQService::ContrastsROPECA$new(mod$modelDF,
 Contr,
 modelFunction$contrast_fun,
 subject_Id = config$table$hkeysDepth(),
 modelName = modelFunction$model_name)

 contrast$get_linfct()

#> dil.b_vs_a=dilution.a - dilution.b

#> $linfct
#>            (Intercept) dilution.b dilution.c dilution.d dilution.e
#> dilution.a           1          0          0          0          0
#> dilution.b           1          1          0          0          0
#> dilution.c           1          0          1          0          0
#> dilution.d           1          0          0          1          0
#> dilution.e           1          0          0          0          1
#> 
#> $linfct_A
#>            (Intercept) dilution.b dilution.c dilution.d dilution.e
#> dil.b_vs_a           0         -1          0          0          0
#> 

 contrast$subject_Id

#> [1] "protein_Id" "peptide_Id"

 tmp <- contrast$get_contrast(all = TRUE)

#> dil.b_vs_a=dilution.a - dilution.b

#> Joining, by = c("contrast", "protein_Id", "peptide_Id")

 head(tmp)

#> # A tibble: 6 x 18
#> # Groups:   contrast [1]
#>   protein_Id contrast     n n_not_na mad.estimate estimate statistic
#>   <chr>      <chr>    <int>    <int>        <dbl>    <dbl>     <dbl>
#> 1 CON__P027~ dil.b_v~    33       33       0.0984 -8.81e-4  -0.00386
#> 2 CON__P349~ dil.b_v~    13       13       0.0667  1.16e-2   0.115  
#> 3 sp|O15511~ dil.b_v~     4        4       0.117   5.30e-2   0.399  
#> 4 sp|O43660~ dil.b_v~     7        7       0.225  -1.51e-2  -0.0427 
#> 5 sp|O95373~ dil.b_v~    26       26       0.194  -2.20e-2  -0.104  
#> 6 sp|P00505~ dil.b_v~    12       12       0.168   5.25e-2   0.168  
#> # ... with 11 more variables: median.p.scaled <dbl>, c1_name <chr>, c1 <dbl>,
#> #   c2_name <chr>, c2 <dbl>, median.p.value <dbl>,
#> #   beta.based.significance <dbl>, n.beta <dbl>, isSingular <lgl>,
#> #   FDR.beta.based.significance <dbl>, FDR.median.p.value <dbl>