compute contrasts based on peptide fold changes - compute median of fold change for each protein. peptide fold change is comptuted based on interaction average

compute contrasts based on peptide fold changes - compute median of fold change for each protein. peptide fold change is comptuted based on interaction average

get_imputed_contrasts(data, config, contrasts)

Arguments

data
data.frame

config
AnalysisConfiguration

contrast
list of contrasts

See also

Examples

library(LFQService)
library(tidyverse)
bb <- LFQServiceData::skylinePRMSampleData_A
configur <- bb$config_f()
data <- bb$analysis(bb$data, configur)

#> creating sampleName

#> completing cases

configur$parameter$qVal_individual_threshold <- 0.01
data <- LFQService::removeLarge_Q_Values(data, configur)
data <- complete_cases(data, configur)

#> completing cases

Contrasts <- c("TimeT168vsT2" = "TimeT168 - TimeT2","TimeT168vsT24" = "TimeT168 - TimeT24" )
get_imputed_contrasts(data, configur, Contrasts)

#> completing cases

#> TimeT168vsT2=TimeT168 - TimeT2

#> TimeT168vsT24=TimeT168 - TimeT24

#> # A tibble: 74 x 8
#>    contrast protein_Id c1_name c2_name     c1     c2 estimate_median
#>    <chr>    <chr>      <chr>   <chr>    <dbl>  <dbl>           <dbl>
#>  1 TimeT16~ CiRT stan~ TimeT1~ TimeT2  9.40e6 1.35e7       -3343250.
#>  2 TimeT16~ iRT-C18 S~ TimeT1~ TimeT2  2.56e7 2.70e7        -457289.
#>  3 TimeT16~ sp|O60656~ TimeT1~ TimeT2  4.29e5 4.50e5          11128.
#>  4 TimeT16~ sp|P02768~ TimeT1~ TimeT2  9.87e8 7.58e8      176831403.
#>  5 TimeT16~ sp|P05177~ TimeT1~ TimeT2  8.32e5 1.04e6           7307.
#>  6 TimeT16~ sp|P09488~ TimeT1~ TimeT2  5.46e4 1.36e5         -96524.
#>  7 TimeT16~ sp|P10632~ TimeT1~ TimeT2  3.52e5 2.26e6       -1893009.
#>  8 TimeT16~ sp|P11509~ TimeT1~ TimeT2  1.34e5 8.68e5        -756395.
#>  9 TimeT16~ sp|P11712~ TimeT1~ TimeT2  3.71e6 6.35e6       -2632496.
#> 10 TimeT16~ sp|P16662~ TimeT1~ TimeT2  5.03e6 7.42e6       -2334678.
#> # ... with 64 more rows, and 1 more variable: estimate_mad <dbl>