compute per group averages and impute values should generalize at some stage

compute per group averages and impute values should generalize at some stage

missigness_impute_factors_interactions(
  pdata,
  config,
  probs = 0.1,
  value = c("nrReplicates", "nrMeasured", "meanArea", "imputed"),
  add.prefix = FALSE,
  global = TRUE
)

Arguments

pdata
data.frame

config
AnalysisConfiguration

probs
quantile to take average from (default 0.1)

value
use default

add.prefix
use default

See also

Examples

bb <- LFQServiceData::skylinePRMSampleData_A
configur <- bb$config_f()
data <- bb$analysis(bb$data, configur)

#> creating sampleName

#> completing cases

configur$parameter$qVal_individual_threshold <- 0.01
xx <- LFQService::removeLarge_Q_Values(data, configur)
xx <- complete_cases(xx, configur)

#> completing cases

res <- missigness_impute_factors_interactions(xx, configur)

#> completing cases

head(res)

#> # A tibble: 6 x 14
#>   value protein_Id peptide_Id precursor_Id fragment_Id Isotope.Label.T~ TimeT0
#>   <chr> <chr>      <chr>      <chr>        <chr>       <chr>             <int>
#> 1 nrRe~ CiRT stan~ ATAGDTHLG~ ATAGDTHLGGE~ ATAGDTHLGG~ light                 3
#> 2 nrRe~ CiRT stan~ ATAGDTHLG~ ATAGDTHLGGE~ ATAGDTHLGG~ light                 3
#> 3 nrRe~ CiRT stan~ ATAGDTHLG~ ATAGDTHLGGE~ ATAGDTHLGG~ light                 3
#> 4 nrRe~ CiRT stan~ ATAGDTHLG~ ATAGDTHLGGE~ ATAGDTHLGG~ light                 3
#> 5 nrRe~ CiRT stan~ ATAGDTHLG~ ATAGDTHLGGE~ ATAGDTHLGG~ light                 3
#> 6 nrRe~ CiRT stan~ ATAGDTHLG~ ATAGDTHLGGE~ ATAGDTHLGG~ light                 3
#> # ... with 7 more variables: TimeT168 <int>, TimeT2 <int>, TimeT24 <int>,
#> #   TimeT240 <int>, TimeT336 <int>, TimeT72 <int>, TimeT8 <int>

res <- missigness_impute_factors_interactions(xx, configur, value = "imputed")

#> completing cases

head(res)

#> # A tibble: 6 x 14
#>   value protein_Id peptide_Id precursor_Id fragment_Id Isotope.Label.T~ TimeT0
#>   <chr> <chr>      <chr>      <chr>        <chr>       <chr>             <dbl>
#> 1 impu~ CiRT stan~ ATAGDTHLG~ ATAGDTHLGGE~ ATAGDTHLGG~ light            1.81e6
#> 2 impu~ CiRT stan~ ATAGDTHLG~ ATAGDTHLGGE~ ATAGDTHLGG~ light            1.26e6
#> 3 impu~ CiRT stan~ ATAGDTHLG~ ATAGDTHLGGE~ ATAGDTHLGG~ light            5.01e5
#> 4 impu~ CiRT stan~ ATAGDTHLG~ ATAGDTHLGGE~ ATAGDTHLGG~ light            3.95e5
#> 5 impu~ CiRT stan~ ATAGDTHLG~ ATAGDTHLGGE~ ATAGDTHLGG~ light            6.72e6
#> 6 impu~ CiRT stan~ ATAGDTHLG~ ATAGDTHLGGE~ ATAGDTHLGG~ light            2.39e6
#> # ... with 7 more variables: TimeT168 <dbl>, TimeT2 <dbl>, TimeT24 <dbl>,
#> #   TimeT240 <dbl>, TimeT336 <dbl>, TimeT72 <dbl>, TimeT8 <dbl>

dim(res)

#> [1] 927  14

dim(dplyr::distinct(res[,1:6]))

#> [1] 927   6

fun <- missigness_impute_factors_interactions(xx, configur, value = "nrMeasured")

#> completing cases