summarize - compute anova and extract model coefficients generated by `model_analyse`

summarize - compute anova and extract model coefficients generated by `model_analyse`

model_analyse_summarize(
  modelProteinF,
  subject_Id = "protein_Id",
  modelName = "Model"
)

See also

Examples

rm(list = ls())
library(LFQService)
D <- LFQService::ionstar$normalized()
D$data <- dplyr::filter(D$data ,protein_Id %in% sample(protein_Id, 10))
modelName <- "f_condtion_r_peptide"
formula_randomPeptide <-
  make_custom_model_lmer("transformedIntensity  ~ dilution. + (1 | peptide_Id)",
   model_name = modelName)
pepIntensity <- D$data
config <- D$config
config$table$hkeysDepth()

#> [1] "protein_Id"

modellingResult <- LFQService:::model_analyse(
 pepIntensity,
 formula_randomPeptide,
 modelName = modelName,
 subject_Id = config$table$hkeysDepth())

#> Joining, by = "protein_Id"

names(modellingResult)

#> [1] "modelProtein" "modelName"   

tmp <- model_analyse_summarize(modellingResult$modelProtein,
subject_Id = config$table$hkeysDepth(),
 modelName = modelName)
head(tmp$Model_Coeff)

#> # A tibble: 6 x 9
#> # Groups:   protein_Id [2]
#>   protein_Id isSingular nrcoef row.names.x. Estimate Std..Error     df t.value
#>   <chr>      <lgl>       <int> <chr>           <dbl>      <dbl>  <dbl>   <dbl>
#> 1 sp|Q58FF8~ FALSE           5 (Intercept)    4.28       0.659    9.04   6.49 
#> 2 sp|Q58FF8~ FALSE           5 dilution.b     0.0421     0.0472 186.     0.892
#> 3 sp|Q58FF8~ FALSE           5 dilution.c    -0.0539     0.0472 186.    -1.14 
#> 4 sp|Q58FF8~ FALSE           5 dilution.d    -0.104      0.0472 186.    -2.22 
#> 5 sp|Q58FF8~ FALSE           5 dilution.e    -0.0818     0.0472 186.    -1.73 
#> 6 sp|P19367~ FALSE           5 (Intercept)   -0.771      0.191   18.5   -4.04 
#> # ... with 1 more variable: Pr...t.. <dbl>

head(tmp$Model_Anova)

#> # A tibble: 6 x 10
#> # Groups:   protein_Id [6]
#>   protein_Id isSingular nrcoef rownames.x.  Sum.Sq Mean.Sq NumDF DenDF F.value
#>   <chr>      <lgl>       <int> <chr>         <dbl>   <dbl> <int> <dbl>   <dbl>
#> 1 sp|Q58FF8~ FALSE           5 dilution.    0.577   0.144      4  186.   3.24 
#> 2 sp|P19367~ FALSE           5 dilution.    0.917   0.229      4  252.   1.51 
#> 3 sp|Q07020~ FALSE           5 dilution.    0.219   0.0549     4  129.   1.87 
#> 4 CON__Q3SZ~ FALSE           5 dilution.    0.0643  0.0161     4  115.   0.488
#> 5 sp|Q9BY44~ FALSE           5 dilution.    1.04    0.261      4  258.   1.80 
#> 6 sp|P0C0V0~ FALSE           5 dilution.   68.3    17.1        4  191. 180.   
#> # ... with 1 more variable: Pr..F. <dbl>