pivot model contrasts matrix to wide format produced by `contrasts_linfct` and ...

pivot model contrasts matrix to wide format produced by `contrasts_linfct` and ...

pivot_model_contrasts_2_Wide(
  modelWithInteractionsContrasts,
  subject_Id = "protein_Id",
  columns = c("estimate", "p.value", "p.value.adjusted"),
  contrast = "lhs"
)

See also

Examples

dd <- LFQServiceData::factor_levelContrasts
head(dd)

#> # A tibble: 6 x 14
#>   Compound lhs   sigma    df estimate std.error statistic p.value conf.low
#>   <chr>    <chr> <dbl> <int>    <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#> 1 4.85_64~ Mort~ 0.987   115   0.128      0.193     0.659   0.511   -0.256
#> 2 4.85_64~ Inte~ 0.987   115   0.0384     0.183     0.210   0.834   -0.324
#> 3 4.85_64~ NRSN~ 0.987   115   0.173      0.222     0.777   0.439   -0.268
#> 4 4.85_64~ NRSN~ 0.987   115  -0.0471     0.222    -0.212   0.833   -0.487
#> 5 4.85_64~ NRSN~ 0.987   115  -0.220      0.221    -0.993   0.323   -0.658
#> 6 5.05_63~ Mort~ 0.719   115   0.104      0.141     0.741   0.460   -0.175
#> # ... with 5 more variables: conf.high <dbl>, isSingular <lgl>,
#> #   sigma.model <dbl>, df.residual.model <dbl>, p.value.adjusted <dbl>

tmp <- pivot_model_contrasts_2_Wide(dd, subject_Id = "Compound")
tmp

#> # A tibble: 30 x 17
#>    Compound isSingular `estimate.Inter~ `estimate.Morta~ `estimate.NRSNR~
#>    <chr>    <lgl>                 <dbl>            <dbl>            <dbl>
#>  1 4.85_64~ FALSE               0.0384            0.128            0.173 
#>  2 5.05_63~ FALSE               0.00972           0.104            0.118 
#>  3 5.11_36~ FALSE               0.245            -0.278            0.367 
#>  4 5.12_26~ FALSE               0.160            -0.431            0.0772
#>  5 5.12_30~ FALSE               0.370            -0.253            0.303 
#>  6 5.12_47~ FALSE              -0.00471           0.436           -0.0256
#>  7 5.13_36~ FALSE              -0.626            -0.255           -0.403 
#>  8 5.16_16~ FALSE               0.236            -0.283           -0.346 
#>  9 5.16_24~ FALSE               0.176            -0.0713          -0.270 
#> 10 5.16_27~ FALSE              -0.0342           -0.132            0.0999
#> # ... with 20 more rows, and 12 more variables: `estimate.NRSNRS_3 -
#> #   NRSNRS_5` <dbl>, `estimate.NRSNRS_4 - NRSNRS_5` <dbl>,
#> #   `p.value.InterventionCTRL - InterventionINT` <dbl>, `p.value.MortalityM0 -
#> #   MortalityM1` <dbl>, `p.value.NRSNRS_3 - NRSNRS_4` <dbl>, `p.value.NRSNRS_3
#> #   - NRSNRS_5` <dbl>, `p.value.NRSNRS_4 - NRSNRS_5` <dbl>,
#> #   `p.value.adjusted.InterventionCTRL - InterventionINT` <dbl>,
#> #   `p.value.adjusted.MortalityM0 - MortalityM1` <dbl>,
#> #   `p.value.adjusted.NRSNRS_3 - NRSNRS_4` <dbl>, `p.value.adjusted.NRSNRS_3 -
#> #   NRSNRS_5` <dbl>, `p.value.adjusted.NRSNRS_4 - NRSNRS_5` <dbl>