read from folder or zip with the content of the MQ txt folder.

read from folder or zip with the content of the MQ txt folder.

tidyMQ_Peptides_Config(
  MQtxtfolder,
  use = c("peptides", "modificationSpecificPeptides"),
  id_extractor = function(df) {     fgcz.gsea.ora::get_UniprotID_from_fasta_header(df,
    idcolumn = "top_protein") },
  remove_rev = TRUE
)

Arguments

MQtxtfolder
path to folder

use
either peptides.txt or modificationSpecificPeptides.txt

id_extractor
function to parse id's

remove_rev
should reverse sequences be removed

See also